libtrx/text: move Text_ChangeText to libtrx
